agent/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
httpx==0.28.1
anthropic==0.49.0
google-cloud-firestore==2.19.0
pydantic-settings==2.7.0
